+++ internal/storage/queries/tasks.sql
@@ -1,20 +1,20 @@
 -- name: GetTask :one
-SELECT * FROM tasks WHERE id = ?;
+SELECT * FROM tasks WHERE id = ? AND user_id = ?;
 
 -- name: GetTasks :many
-SELECT * FROM tasks ORDER BY tasks.created_at DESC LIMIT ? OFFSET ?;
+SELECT * FROM tasks WHERE user_id = ? ORDER BY tasks.created_at DESC LIMIT ? OFFSET ?;
 
 -- name: CreateTask :one
-INSERT INTO tasks (id, title, description, completed, created_at, updated_at)
-VALUES (?, ?, ?, ?, ?, ?)
+INSERT INTO tasks (id, title, description, completed, created_at, updated_at, user_id)
+VALUES (?, ?, ?, ?, ?, ?, ?)
 RETURNING *;
 
 -- name: DeleteTask :one
-DELETE FROM tasks WHERE id = ?
+DELETE FROM tasks WHERE id = ? AND user_id = ?
 RETURNING id;
 
 -- name: UpdateTask :one
 UPDATE tasks
 SET title = ?, description = ?, completed = ?, updated_at = ?
-WHERE id = ?
-RETURNING *
+WHERE id = ? AND user_id = ?
+RETURNING *;
